+++ Changelog.md
@@ -1,3 +1,9 @@
-## 1.3.3
+## 0.1.3
 - added filter for pre-update token to prevent non-effect updates triggering the main workflow
 - fixed CreateActiveEffect function that previously saved token ID to the aura map, rather than updating on a single applied token.
+
+## 0.1.4
+- Added option for Alignment/Type checks before application
+- Added optional save before application
+- More bug fixes to prevent multiple application(check for auras before running workflow on token create)
+- Auras no-longer apply when a token is at 0 hp (optional toggle in the settings)